Cloudflare captcha handling
In combination with Referer Control for Firefox or Chrome this can successfully handle Cloudflare protection.
If script find out there is such protection, it get us link to NoScript Recaptcha and wait for text to be pasted (and then Enter pressed).
But Recaptcha uses referer to check what site actually uses it, so we temporary need to spoof it to e621.net.
And here comes Referer Control. With site filter "https://www.google.com/recaptcha/*" and settings "To" and "Custom" and "https://e621.net" we can successfully solve a captcha and pass a solution text to e621dl.
This should be done only once a session.